+++ docs/src/app/pages/docs/components/app-components-qrcode/app-components-qrcode.tsx
@@ -72,13 +72,13 @@ export class AppComponentsQRCode {
 <h2 id="app-components-qrcode-getting-started">Getting started</h2>
 <p>To create easily your PWA presentation and to enjoy all the options, I suggest you to create your slides using the CLI as described in the <a href="/docs/introduction">Getting started chapter</a>.</p>
 <p>Doing so you will use the <a href="https://deckdeckgo.com">DeckDeckGo</a> starter kit which already includes this Web Component.</p>
-<p>However, if you are looking to use this Web Component as a standalone component, to add a QR codes to your web applications, it could be use directly in your project from a CDN, using a simple script include, or could be installed from <a href="https://www.npmjs.com/package/deckdeckgo-qrcode">npm</a>.</p>
+<p>However, if you are looking to use this Web Component as a standalone component, to add a QR codes to your web applications, it could be use directly in your project from a CDN, using a simple script include, or could be installed from <a href="https://www.npmjs.com/package/@deckdeckgo/qrcode">npm</a>.</p>
 <h3 id="app-components-qrcode-using-deckdeckgo-qr-code-from-a-cdn">Using DeckDeckGo QR Code from a CDN</h3>
 <p>It&#39;s recommended to use <a href="https://unpkg.com/">unpkg</a> to use the <a href="https://deckdeckgo.com">DeckDeckGo</a> Code from a CDN. To do so, add the following include script in the main HTML file of your project:</p>
 <deckgo-highlight-code language="javascript">
       <code slot="code">&lt;script src=&quot;https:&#47;&#47;unpkg.com&#47;deckdeckgo-qrcode@latest&#47;dist&#47;deckdeckgo-qrcode.js&quot;&gt;&lt;&#47;script&gt;</code>
     </deckgo-highlight-code><h3 id="app-components-qrcode-install-deckdeckgo-qr-code-from-npm">Install DeckDeckGo QR Code from NPM</h3>
-<p>Install <a href="https://deckdeckgo.com">DeckDeckGo</a> - QR Code in your project from <a href="https://www.npmjs.com/package/deckdeckgo-qrcode">npm</a> using the following command:</p>
+<p>Install <a href="https://deckdeckgo.com">DeckDeckGo</a> - QR Code in your project from <a href="https://www.npmjs.com/package/@deckdeckgo/qrcode">npm</a> using the following command:</p>
 <deckgo-highlight-code language="bash">
       <code slot="code">npm install deckdeckgo-qrcode</code>
     </deckgo-highlight-code><h3 id="app-components-qrcode-framework-integration">Framework integration</h3>