+++ docs/src/global/app.ts
@@ -1,7 +1,7 @@
 import '@ionic/core';
 
-import 'deckdeckgo';
-import 'deckdeckgo-highlight-code';
-import 'deckdeckgo-charts';
-import 'deckdeckgo-qrcode';
-import 'deckdeckgo-inline-editor';
+import '@deckdeckgo/core';
+import '@deckdeckgo/highlight-code';
+import '@deckdeckgo/charts';
+import '@deckdeckgo/qrcode';
+import '@deckdeckgo/inline-editor';