+++ studio/src/global/app-dev.ts
@@ -1,11 +1,11 @@
 import '@ionic/core';
 
-import 'deckdeckgo';
-import 'deckdeckgo-inline-editor';
-import 'deckdeckgo-remote';
-import 'deckdeckgo-qrcode';
-import 'deckdeckgo-highlight-code';
-import 'deckdeckgo-lazy-img';
+import '@deckdeckgo/core';
+import '@deckdeckgo/inline-editor';
+import '@deckdeckgo/remote';
+import '@deckdeckgo/qrcode';
+import '@deckdeckgo/highlight-code';
+import '@deckdeckgo/lazy-img';
 
 import {setupConfig} from '../app/services/core/environment/environment-config';